app/requirements.txt
Django
channels
channels_redis
djangorestframework
